Incorporate rule summaries in team menu buttons
